first iteration of cleanup. minor bugfix in triple vertex output
