+++ immunopepper/utils.py
@@ -8,12 +8,7 @@ import logging
 
 from .constant import NOT_EXIST
 from .immuno_nametuple import Coord
-from .immuno_nametuple import Flag
 from .immuno_nametuple import Idx
-from .immuno_nametuple import OutputBackground
-from .immuno_nametuple import Peptide
-from .immuno_nametuple import ReadingFrameTuple
-
 
 def to_adj_list(adj_matrix):
     """
@@ -73,14 +68,6 @@ def leq_strand(coord1, coord2, strand):
         return coord1 >= coord2
 
 
-def complementary_seq(dna_seq):
-    """ Yields the complementary DNA sequence
-    Only convert the character in comp_dict.
-    Otherwise remain the same.
-    """
-    comp_dict = {"A": "T", "T": "A", "C": "G", "G": "C"}
-    return "".join([comp_dict[nuc] if nuc in comp_dict else nuc for nuc in dna_seq])
-
 
 def encode_chromosome(in_num):
     """
@@ -88,59 +75,6 @@ def encode_chromosome(in_num):
     """
     convert_dict = {23: "X", 24: "Y", 25: "MT"}
     return convert_dict[in_num] if in_num in convert_dict else str(in_num)
-
-
-def translate_dna_to_peptide(dna_str):
-    """ Translate a DNA sequence encoding a peptide to amino-acid sequence via RNA.
-
-    If 'N' is included in input dna, 'X' will be outputted since 'N' represents
-    uncertainty. Also will output a flag indicating if has stop codon.
-
-    Parameters
-    ----------
-    dna_str: str or List(str). dna string to be translated.
-
-    Returns
-    -------
-    aa_str: translated peptide
-    has_stop_codon: Indicator for showing if the input dna contains stop codon
-
-    """
-    codontable = {
-        'ATA': 'I', 'ATC': 'I', 'ATT': 'I', 'ATG': 'M',
-        'ACA': 'T', 'ACC': 'T', 'ACG': 'T', 'ACT': 'T',
-        'AAC': 'N', 'AAT': 'N', 'AAA': 'K', 'AAG': 'K',
-        'AGC': 'S', 'AGT': 'S', 'AGA': 'R', 'AGG': 'R',
-        'CTA': 'L', 'CTC': 'L', 'CTG': 'L', 'CTT': 'L',
-        'CCA': 'P', 'CCC': 'P', 'CCG': 'P', 'CCT': 'P',
-        'CAC': 'H', 'CAT': 'H', 'CAA': 'Q', 'CAG': 'Q',
-        'CGA': 'R', 'CGC': 'R', 'CGG': 'R', 'CGT': 'R',
-        'GTA': 'V', 'GTC': 'V', 'GTG': 'V', 'GTT': 'V',
-        'GCA': 'A', 'GCC': 'A', 'GCG': 'A', 'GCT': 'A',
-        'GAC': 'D', 'GAT': 'D', 'GAA': 'E', 'GAG': 'E',
-        'GGA': 'G', 'GGC': 'G', 'GGG': 'G', 'GGT': 'G',
-        'TCA': 'S', 'TCC': 'S', 'TCG': 'S', 'TCT': 'S',
-        'TTC': 'F', 'TTT': 'F', 'TTA': 'L', 'TTG': 'L',
-        'TAC': 'Y', 'TAT': 'Y', 'TAA': '_', 'TAG': '_',
-        'TGC': 'C', 'TGT': 'C', 'TGA': '_', 'TGG': 'W'
-    }
-    dna_str = dna_str.upper()
-    has_stop_codon = False
-    aa_str = []
-    for idx in range(0, len(dna_str), 3):
-        codon = dna_str[idx:idx + 3]
-        if len(codon) < 3:
-            break
-        if 'N' in codon:
-            aa_str.append('X')
-        else:
-            if codontable[codon] == '_':
-                has_stop_codon = True
-                return ''.join(aa_str), has_stop_codon
-            else:
-                aa_str.append(codontable[codon])
-
-    return ''.join(aa_str), has_stop_codon
 
 
 def get_sub_mut_dna(background_seq, coord, variant_comb, somatic_mutation_sub_dict, strand):
@@ -199,163 +133,6 @@ def get_sub_mut_dna(background_seq, coord, variant_comb, somatic_mutation_sub_di
             sub_dna = sub_dna[:pos] + mut_base + sub_dna[pos+1:]
     return sub_dna
 
-
-def cross_peptide_result(read_frame, strand, variant_comb, somatic_mutation_sub_dict, ref_mut_seq, peptide_accept_coord):
-    """ Get translated peptide from the given exon pairs.
-
-    Parameters
-    ----------
-    read_frame: NamedTuple. (read_start_codon, read_stop_codon, emitting_frame)
-    strand: str. '+' or '-'
-    variant_comb: List(int).
-    somatic_mutation_sub_dict: Dict. variant position -> variant details.
-    ref_mut_seq: Dict.['ref', 'background'] -> List(str)
-    peptide_accept_coord: The start and end position of next vertex. Positions of the first vertex
-        are already given in read_frame.
-
-    Returns
-    -------
-    peptide: NamedTuple Peptide. has attribute ['ref', 'mut']. contain the output peptide
-        translated from reference sequence and mutated sequence.
-    coord: NamedTuple Coord. has attribute ['start_v1', 'stop_v1', 'start_v2', 'stop_v2']
-        contains the true four position of exon pairs (after considering read framee)
-        that outputs the peptide.
-    flag: NamedTuple Flag. has attribute ['has_stop', 'is_isolated']
-    next_reading_frame: Tuple. The reading frame to be propogated to the next vertex.
-
-    """
-    cds_left_modi, cds_right_modi, emitting_frame = read_frame[0], read_frame[1], read_frame[2]
-    next_emitting_frame = (peptide_accept_coord[1] - peptide_accept_coord[0] + emitting_frame) % 3
-    start_v1 = cds_left_modi
-    stop_v1 = cds_right_modi
-
-    #                                       |next_start_v1  |
-    # |      v1           | |    v2                         |
-    # -----[emitting_frame] [accepting_frame]-------
-    # emitting_frame + accepting_frame = 3
-    accepting_frame = (3 - emitting_frame) % 3
-
-    if somatic_mutation_sub_dict:  # exist maf dictionary, so we use germline mutation-applied seq as the background seq
-        ref_seq = ref_mut_seq['background']
-    else:
-        ref_seq = ref_mut_seq['ref']
-    mut_seq = ref_mut_seq['background']
-    # python is 0-based while gene annotation file(.gtf) is one based
-    # so we need to do a little modification
-    if strand == "+":
-        start_v2 = peptide_accept_coord[0]
-        stop_v2 = max(start_v2, peptide_accept_coord[1] - next_emitting_frame)
-        coord = Coord(start_v1, stop_v1, start_v2, stop_v2)
-        peptide_dna_str_mut = get_sub_mut_dna(mut_seq, coord, variant_comb, somatic_mutation_sub_dict, strand)
-        peptide_dna_str_ref = ref_seq[start_v1:stop_v1] + ref_seq[start_v2:stop_v2]
-        next_start_v1 = min(start_v2 + accepting_frame, peptide_accept_coord[1])
-        next_stop_v1 = peptide_accept_coord[1]
-    else:  # strand == "-"
-        stop_v2 = peptide_accept_coord[1]
-        start_v2 = min(stop_v2, peptide_accept_coord[0] + next_emitting_frame)
-        coord = Coord(start_v1, stop_v1, start_v2, stop_v2)
-        peptide_dna_str_mut = complementary_seq(get_sub_mut_dna(mut_seq, coord, variant_comb, somatic_mutation_sub_dict, strand))
-        peptide_dna_str_ref = complementary_seq(ref_seq[start_v1:stop_v1][::-1] + ref_seq[start_v2:stop_v2][::-1])
-        next_start_v1 = peptide_accept_coord[0]
-        next_stop_v1 = max(stop_v2 - accepting_frame, peptide_accept_coord[0])
-
-    next_reading_frame = ReadingFrameTuple(next_start_v1, next_stop_v1, next_emitting_frame)
-    assert (len(peptide_dna_str_mut) == len(peptide_dna_str_ref))
-    # if len(peptide_dna_str_mut) % 3 != 0:
-    #     print("Applied mutations have changed the length of the DNA fragment - no longer divisible by 3")
-    peptide_mut, mut_has_stop_codon = translate_dna_to_peptide(peptide_dna_str_mut)
-    peptide_ref, ref_has_stop_codon = translate_dna_to_peptide(peptide_dna_str_ref)
-
-    # if the stop codon appears before translating the second exon, mark 'single'
-    is_isolated = False
-    if len(peptide_mut)*3 <= abs(stop_v1 - start_v1) + 1:
-        is_isolated = True
-        jpos = 0.0
-    else:
-        jpos = float(stop_v1 - start_v1) / 3.0
-    peptide = Peptide(peptide_mut, peptide_ref)
-    coord = Coord(start_v1, stop_v1, start_v2, stop_v2)
-    flag = Flag(mut_has_stop_codon, is_isolated)
-    return peptide, coord, flag, next_reading_frame
-
-
-def isolated_peptide_result(read_frame, strand, variant_comb, somatic_mutation_sub_dict, ref_mut_seq):
-    """ Deal with translating isolated cds, almost the same as cross_peptide_result
-
-    Parameters
-    ----------
-    read_frame: Tuple. (read_start_codon, read_stop_codon, emitting_frame)
-    strand: str. '+' or '-'
-    variant_comb: List(int).
-    somatic_mutation_sub_dict: Dict. variant position -> variant details.
-    ref_mut_seq: Dict.['ref', 'background'] -> List(str)
-
-    Returns
-    -------
-    peptide: NamedTuple. has attribute ['ref', 'mut']. contain the output peptide
-        translated from reference sequence and mutated sequence.
-    coord: NamedTuple. has attribute ['start_v1', 'stop_v1', 'start_v2', 'stop_v2']
-        contains the true two position of exon pairs (after considering read framee)
-        that outputs the peptide. 'start_v2', 'stop_v2' is set to be NOT_EXIST.
-    flag: NamedTuple. has attribute ['has_stop', 'is_isolated']
-
-    """
-
-    start_v1, stop_v1, emitting_frame = read_frame.cds_left_modi,read_frame.cds_right_modi,read_frame.read_phase
-    start_v2 = NOT_EXIST
-    stop_v2 = NOT_EXIST
-
-    if somatic_mutation_sub_dict:  # exist maf dictionary, so we use germline mutation-applied seq as the background seq
-        ref_seq = ref_mut_seq['background']
-    else:
-        ref_seq = ref_mut_seq['ref']
-    mut_seq = ref_mut_seq['background']
-
-    if strand == '+':
-        coord = Coord(start_v1, stop_v1, start_v2, stop_v2)
-        peptide_dna_str_mut = get_sub_mut_dna(mut_seq, coord, variant_comb, somatic_mutation_sub_dict, strand)
-        peptide_dna_str_ref = ref_seq[start_v1:stop_v1]
-    else:
-        coord = Coord(start_v1, stop_v1, start_v2, stop_v2)
-        peptide_dna_str_mut = complementary_seq(get_sub_mut_dna(mut_seq, coord, variant_comb, somatic_mutation_sub_dict, strand))
-        peptide_dna_str_ref = complementary_seq(ref_seq[start_v1:stop_v1][::-1])
-
-    peptide_mut, mut_has_stop_codon = translate_dna_to_peptide(peptide_dna_str_mut)
-    peptide_ref, ref_has_stop_codon = translate_dna_to_peptide(peptide_dna_str_ref)
-
-    is_isolated = True
-
-    peptide = Peptide(peptide_mut,peptide_ref)
-    coord = Coord(start_v1, stop_v1, start_v2, stop_v2)
-    flag = Flag(mut_has_stop_codon,is_isolated)
-
-    return peptide, coord, flag
-
-
-def get_peptide_result(simple_meta_data, strand, variant_comb, somatic_mutation_sub_dict, ref_mut_seq):
-    if somatic_mutation_sub_dict:  # exist maf dictionary, so we use germline mutation-applied seq as the background seq
-        ref_seq = ref_mut_seq['background']
-    else:
-        ref_seq = ref_mut_seq['ref']
-    mut_seq = ref_mut_seq['background']
-    modi_coord = simple_meta_data.modified_exons_coord
-    if strand == "+":
-        peptide_dna_str_mut = get_sub_mut_dna(mut_seq, modi_coord, variant_comb, somatic_mutation_sub_dict, strand)
-        peptide_dna_str_ref = get_sub_mut_dna(ref_seq, modi_coord, NOT_EXIST, somatic_mutation_sub_dict, strand)
-    else:  # strand == "-"
-        peptide_dna_str_mut = complementary_seq(get_sub_mut_dna(mut_seq, modi_coord, variant_comb, somatic_mutation_sub_dict, strand))
-        peptide_dna_str_ref = complementary_seq(get_sub_mut_dna(ref_seq, modi_coord, NOT_EXIST, somatic_mutation_sub_dict, strand))
-    peptide_mut, mut_has_stop_codon = translate_dna_to_peptide(peptide_dna_str_mut)
-    peptide_ref, ref_has_stop_codon = translate_dna_to_peptide(peptide_dna_str_ref)
-
-    # if the stop codon appears before translating the second exon, mark 'single'
-    if modi_coord.start_v2 == NOT_EXIST or len(peptide_mut)*3 <= abs(modi_coord.stop_v1 - modi_coord.start_v1) + 1:
-        is_isolated = True
-    else:
-        is_isolated = False
-    peptide = Peptide(peptide_mut, peptide_ref)
-    flag = Flag(mut_has_stop_codon, is_isolated)
-    return peptide,flag
 
 
 def get_size_factor(strains, lib_file_path):
@@ -422,22 +199,6 @@ def is_isolated_cds(gene, idx):
         return False
 
     return np.sum(gene.splicegraph.edges[:, idx]) == 0
-
-
-def is_in_junction_list(splicegraph, vertex_id_list, strand, junction_list):
-    """Check if the intron is in concerned junction list"""
-
-    if NOT_EXIST in vertex_id_list or junction_list is None:
-        return NOT_EXIST
-
-    vertex_coord_list = splicegraph.vertices
-    if strand == '-':
-        vertex_id_list = vertex_id_list[::-1] # in negative strand case, the vertex id list is [5,3] instead of [3,5]
-    for i in range(len(vertex_id_list) - 1):
-        junc_comb = ':'.join([str(vertex_coord_list[1, vertex_id_list[i]]),  str(vertex_coord_list[0, vertex_id_list[i+1]]), strand])
-        if junc_comb in junction_list:
-            return 1
-    return 0
 
 
 def get_exon_expr(gene, vstart, vstop, countinfo, Idx, seg_counts=None):
